Cambios en hooks para obtener informacion de endpoints. Refactorización de algunos componentes.
