+++ src/utils/hooks/useWizelineGetEndpoints.js
@@ -4,10 +4,10 @@ import { useLatestAPI } from "./useLatestAPI";
 import PropTypes from "prop-types";
 
 export function useWizelineGetEndpoints(url) {
-  const { ref: apiRef, isLoading: isApiMetadataLoading } = useLatestAPI();
+  const { ref: apiRef, loadingResponse: isApiMetadataLoading } = useLatestAPI();
   const [response, setResponseStatus] = useState(() => ({
     data: {},
-    isLoading: true,
+    loadingResponse: true,
   }));
 
   useEffect(() => {
@@ -20,14 +20,14 @@ export function useWizelineGetEndpoints(url) {
     async function getFromEndpoint() {
       let apiParameters = url.replace("{apiRef}", apiRef);
       try {
-        setResponseStatus({ data: { results: [] }, isLoading: true });
+        setResponseStatus({ data: { results: [] }, loadingResponse: true });
         const response = await fetch(`${API_BASE_URL}${apiParameters}`, {
           signal: controller.signal,
         });
         const data = await response.json()
-        setResponseStatus({ data, isLoading: false });
+        setResponseStatus({ data, loadingResponse: false });
       } catch (err) {
-        setResponseStatus({ data: { results: [] }, isLoading: false });
+        setResponseStatus({ data: { results: [] }, loadingResponse: false });
       }
     }
